More flrc testing and optimisation
Double sends at 2k raw seem to be working more reliably with the prototype
boards now that continuous receives are implemented. It doesn't seem to work
well on the breadboard prototypes, especially the larger one. This might be
worth giving a test flight to see how the LQ holds up