Add FEI flag to telem packets
Added a flag to telem packets to indicate if the receiver has measured
the FEI and applied a correction. Use the flag to control if the link is
reset to 125Hz when a new connection is established.
Also removed the unimplemented serialData field for now and moved the
crc to the end of the packet.