Limit snr range sent to FC, send extra linkstats packets
Snr to BF has a max of 127 or it wraps.
Add extra linkstats packets when the packet rate changes, so that
BF can get the packet rate from the linkstats instead of estimating
it itself. Partially working, there still seem to be some unexpected
delays as if not all the linkstats packets are getting through